Ability to disable default trivia
